requirements.txt
pdfplumber
google-generativeai
python-dotenv
weasyprint
markdown
mistletoe
xhtml2pdf